Create CWK Issue
hi maxime, iam ahmed, i expect u remembered me in the past three days ago, when u plays with me in pvp, I was never cheating, I was just clicking on the settings to exit because I expected that I will lose, and suddenly there was a problem that needed to be corrected, and I sorry for what happened during it.
Please unblock me, I want to enjoy my childhood game, and I promise you this will not happen like I did. please.